for ecommerce

+++ Backend/src/main/java/com/ecommerce/root/response/ProductResponse.java
@@ -1,6 +1,5 @@
 package com.ecommerce.root.response;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
@@ -13,7 +12,7 @@ public class ProductResponse {
     private String name;
     private String description;
     private String image;
-    private Double price;
+    private Long price;
     private Long quantity;
     @JsonProperty("category_name")
     private String categoryName;
